Fix: file structure and adding more methods feat: adding github actions
